Split package into multiple source files
This is getting big enough that a bit of reorganization seems useful.
